requirements.txt
# Astro Runtime includes the following pre-installed providers packages: https://www.astronomer.io/docs/astro/runtime-image-architecture#provider-packages
apache-airflow-providers-google
# astronomer-cosmos==1.9.0
astronomer-cosmos==1.8.2
astronomer-cosmos[google]
# pydantic